Add round setup logic
Implemented the Round class which is created using a "RoundCreator". overall im happy with how it works, the system makes it really simple and unfrustrating to add additional predefined "rounds". to do next is gamelogic/gamemanager which uses the round definitions to spawn in rounds of enemies as one would expect them to be. then, the game will be 100% functional (not the same as complete)